+++ src/pypi_attestations/_impl.py
@@ -38,6 +38,45 @@ if TYPE_CHECKING:  # pragma: no cover
     from sigstore.verify.policy import VerificationPolicy
 
 
+# List the claims OID supported
+# Source: https://github.com/sigstore/fulcio/blob/main/docs/oid-info.md
+# We only support the extensions from 1.3.6.1.4.1.57264.1.8 to .22.
+# In particular, `1.3.6.1.4.1.57264.1.7 | OtherName SAN` is not supported
+# because we believe this is not used in-the-wild.
+_FULCIO_CLAIMS_OIDS = [
+    # 1.3.6.1.4.1.57264.1.8 | Issuer (V2)
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.8"),
+    # 1.3.6.1.4.1.57264.1.9 | Build Signer URI
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.9"),
+    # 1.3.6.1.4.1.57264.1.10 | Build Signer Digest
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.10"),
+    # 1.3.6.1.4.1.57264.1.11 | Runner Environment
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.11"),
+    # 1.3.6.1.4.1.57264.1.12 | Source Repository URI
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.12"),
+    # 1.3.6.1.4.1.57264.1.13 | Source Repository Digest
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.13"),
+    # 1.3.6.1.4.1.57264.1.14 | Source Repository Ref
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.14"),
+    # 1.3.6.1.4.1.57264.1.15 | Source Repository Identifier
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.15"),
+    # 1.3.6.1.4.1.57264.1.16 | Source Repository Owner URI
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.16"),
+    # 1.3.6.1.4.1.57264.1.17 | Source Repository Owner Identifier
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.17"),
+    # 1.3.6.1.4.1.57264.1.18 | Build Config URI
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.18"),
+    # 1.3.6.1.4.1.57264.1.19 | Build Config Digest
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.19"),
+    # 1.3.6.1.4.1.57264.1.20 | Build Trigger
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.20"),
+    # 1.3.6.1.4.1.57264.1.21 | Run Invocation URI
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.21"),
+    # 1.3.6.1.4.1.57264.1.22 | Source Repository Visibility At Signing
+    x509.ObjectIdentifier("1.3.6.1.4.1.57264.1.22"),
+]
+
+
 class Distribution(BaseModel):
     """Represents a Python package distribution.
 
@@ -166,6 +205,24 @@ class Attestation(BaseModel):
             return Attestation.from_bundle(bundle)
         except ConversionError as e:
             raise AttestationError(str(e))
+
+    @property
+    def certificate_claims(self) -> dict[str, str]:
+        """Return the claims present in the certificate.
+
+        We only return claims present in `_FULCIO_CLAIMS_OIDS`.
+        Values are decoded and returned as strings.
+        """
+        certificate = x509.load_der_x509_certificate(self.verification_material.certificate)
+        claims = {}
+        for extension in certificate.extensions:
+            if extension.oid in _FULCIO_CLAIMS_OIDS:
+                # 1.3.6.1.4.1.57264.1.8 through 1.3.6.1.4.1.57264.1.22 are formatted as DER-encoded
+                # strings; the ASN.1 tag is UTF8String (0x0C) and the tag class is universal.
+                value = extension.value.value
+                claims[extension.oid.dotted_string] = _der_decode_utf8string(value)
+
+        return claims
 
     def verify(
         self,